Renames folder as per intended scope.
