+++ Components/OPx/Implementation/EnvelopeGenerator.hpp
