+++ OSBindings/SDL/SConstruct
@@ -50,8 +50,7 @@ SOURCES += glob.glob('../../Components/AudioToggle/*.cpp')
 SOURCES += glob.glob('../../Components/AY38910/*.cpp')
 SOURCES += glob.glob('../../Components/DiskII/*.cpp')
 SOURCES += glob.glob('../../Components/KonamiSCC/*.cpp')
-SOURCES += glob.glob('../../Components/OPL2/*.cpp')
-SOURCES += glob.glob('../../Components/OPL2/Implementation/*.cpp')
+SOURCES += glob.glob('../../Components/OPLx/*.cpp')
 SOURCES += glob.glob('../../Components/SN76489/*.cpp')
 SOURCES += glob.glob('../../Components/Serial/*.cpp')
